+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/ServiceClients/LanguageServiceClient.cs
@@ -10,7 +10,6 @@ using System.Threading.Tasks;
 using Azure.AI.TextAnalytics.Models;
 using Azure.Core;
 using Azure.Core.Pipeline;
-using static System.Collections.Specialized.BitVector32;
 
 namespace Azure.AI.TextAnalytics.ServiceClients
 {
@@ -2085,6 +2084,258 @@ namespace Azure.AI.TextAnalytics.ServiceClients
                 IDictionary<string, int> idToIndexMap = CreateIdToIndexMap(multiLanguageInput.Documents);
 
                 return new ClassifyDocumentOperation(this, _clientDiagnostics, location, idToIndexMap, options.IncludeStatistics);
+            }
+            catch (Exception e)
+            {
+                scope.Failed(e);
+                throw;
+            }
+        }
+
+        #endregion
+
+        #region Dynamic Classify
+
+        public override Response<ClassificationCategoryCollection> DynamicClassify(
+            string document,
+            IEnumerable<string> categories,
+            string language = default,
+            DynamicClassifyOptions options = default,
+            CancellationToken cancellationToken = default)
+        {
+            Argument.AssertNotNullOrEmpty(document, nameof(document));
+            Argument.AssertNotNullOrEmpty(categories, nameof(categories));
+
+            options ??= new DynamicClassifyOptions();
+
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TextAnalyticsClient)}.{nameof(DynamicClassify)}");
+            scope.AddAttribute("document", document);
+            scope.Start();
+
+            try
+            {
+                MultiLanguageAnalysisInput multiLanguageInput = new();
+                multiLanguageInput.Documents.Add(ConvertToMultiLanguageInput(document, language));
+
+                AnalyzeTextDynamicClassificationInput input = new()
+                {
+                    AnalysisInput = multiLanguageInput,
+                    Parameters = new DynamicClassificationTaskParameters(
+                        options.DisableServiceLogs,
+                        options.ModelVersion,
+                        options.ClassificationType,
+                        categories.ToList())
+                };
+
+                Response<AnalyzeTextTaskResult> result = _languageRestClient.Analyze(
+                    input,
+                    options.IncludeStatistics,
+                    cancellationToken: cancellationToken);
+
+                DynamicClassificationTaskResult taskResult = (DynamicClassificationTaskResult)result.Value;
+                Response response = result.GetRawResponse();
+
+                if (taskResult.Results.Errors.Count > 0)
+                {
+                    // Only one document, so we can ignore the ID and grab the first error message.
+                    TextAnalyticsError error = Transforms.ConvertToError(taskResult.Results.Errors[0].Error);
+                    throw _clientDiagnostics.CreateRequestFailedException(response, new ResponseError(error.ErrorCode.ToString(), error.Message), CreateAdditionalInformation(error));
+                }
+
+                return Response.FromValue(Transforms.ConvertToClassificationCategoryCollection(taskResult.Results.Documents[0]), response);
+            }
+            catch (Exception e)
+            {
+                scope.Failed(e);
+                throw;
+            }
+        }
+
+        public override Response<DynamicClassifyDocumentResultCollection> DynamicClassifyBatch(
+            IEnumerable<string> documents,
+            IEnumerable<string> categories,
+            string language = default,
+            DynamicClassifyOptions options = default,
+            CancellationToken cancellationToken = default)
+        {
+            Argument.AssertNotNullOrEmpty(documents, nameof(documents));
+            Argument.AssertNotNullOrEmpty(categories, nameof(categories));
+            MultiLanguageAnalysisInput documentInputs = ConvertToMultiLanguageInputs(documents, language);
+
+            return DynamicClassifyBatch(documentInputs, categories, options, cancellationToken);
+        }
+
+        public override Response<DynamicClassifyDocumentResultCollection> DynamicClassifyBatch(
+            IEnumerable<TextDocumentInput> documents,
+            IEnumerable<string> categories,
+            DynamicClassifyOptions options = default,
+            CancellationToken cancellationToken = default)
+        {
+            Argument.AssertNotNullOrEmpty(documents, nameof(documents));
+            Argument.AssertNotNullOrEmpty(categories, nameof(categories));
+            MultiLanguageAnalysisInput documentInputs = ConvertToMultiLanguageInputs(documents);
+
+            return DynamicClassifyBatch(documentInputs, categories, options, cancellationToken);
+        }
+
+        public override async Task<Response<ClassificationCategoryCollection>> DynamicClassifyAsync(
+            string document,
+            IEnumerable<string> categories,
+            string language = default,
+            DynamicClassifyOptions options = default,
+            CancellationToken cancellationToken = default)
+        {
+            Argument.AssertNotNullOrEmpty(document, nameof(document));
+            Argument.AssertNotNullOrEmpty(categories, nameof(categories));
+
+            options ??= new DynamicClassifyOptions();
+
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TextAnalyticsClient)}.{nameof(DynamicClassify)}");
+            scope.AddAttribute("document", document);
+            scope.Start();
+
+            try
+            {
+                MultiLanguageAnalysisInput multiLanguageInput = new();
+                multiLanguageInput.Documents.Add(ConvertToMultiLanguageInput(document, language));
+
+                AnalyzeTextDynamicClassificationInput input = new()
+                {
+                    AnalysisInput = multiLanguageInput,
+                    Parameters = new DynamicClassificationTaskParameters(
+                        options.DisableServiceLogs,
+                        options.ModelVersion,
+                        options.ClassificationType,
+                        categories.ToList())
+                };
+
+                Response<AnalyzeTextTaskResult> result = await _languageRestClient.AnalyzeAsync(
+                    input,
+                    options.IncludeStatistics,
+                    cancellationToken: cancellationToken).ConfigureAwait(false);
+
+                DynamicClassificationTaskResult taskResult = (DynamicClassificationTaskResult)result.Value;
+                Response response = result.GetRawResponse();
+
+                if (taskResult.Results.Errors.Count > 0)
+                {
+                    // Only one document, so we can ignore the ID and grab the first error message.
+                    TextAnalyticsError error = Transforms.ConvertToError(taskResult.Results.Errors[0].Error);
+                    throw await _clientDiagnostics.CreateRequestFailedExceptionAsync(response, new ResponseError(error.ErrorCode.ToString(), error.Message), CreateAdditionalInformation(error)).ConfigureAwait(false);
+                }
+
+                return Response.FromValue(Transforms.ConvertToClassificationCategoryCollection(taskResult.Results.Documents[0]), response);
+            }
+            catch (Exception e)
+            {
+                scope.Failed(e);
+                throw;
+            }
+        }
+
+        public override async Task<Response<DynamicClassifyDocumentResultCollection>> DynamicClassifyBatchAsync(
+            IEnumerable<string> documents,
+            IEnumerable<string> categories,
+            string language = default,
+            DynamicClassifyOptions options = default,
+            CancellationToken cancellationToken = default)
+        {
+            Argument.AssertNotNullOrEmpty(documents, nameof(documents));
+            Argument.AssertNotNullOrEmpty(categories, nameof(categories));
+            MultiLanguageAnalysisInput documentInputs = ConvertToMultiLanguageInputs(documents, language);
+
+            return await DynamicClassifyBatchAsync(documentInputs, categories, options, cancellationToken).ConfigureAwait(false);
+        }
+
+        public override async Task<Response<DynamicClassifyDocumentResultCollection>> DynamicClassifyBatchAsync(
+            IEnumerable<TextDocumentInput> documents,
+            IEnumerable<string> categories,
+            DynamicClassifyOptions options = default,
+            CancellationToken cancellationToken = default)
+        {
+            Argument.AssertNotNullOrEmpty(documents, nameof(documents));
+            Argument.AssertNotNullOrEmpty(categories, nameof(categories));
+            MultiLanguageAnalysisInput documentInputs = ConvertToMultiLanguageInputs(documents);
+
+            return await DynamicClassifyBatchAsync(documentInputs, categories, options, cancellationToken).ConfigureAwait(false);
+        }
+
+        private Response<DynamicClassifyDocumentResultCollection> DynamicClassifyBatch(
+            MultiLanguageAnalysisInput multiLanguageInput,
+            IEnumerable<string> categories,
+            DynamicClassifyOptions options,
+            CancellationToken cancellationToken)
+        {
+            options ??= new DynamicClassifyOptions();
+
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TextAnalyticsClient)}.{nameof(DynamicClassifyBatch)}");
+            scope.Start();
+
+            try
+            {
+                AnalyzeTextDynamicClassificationInput input = new()
+                {
+                    AnalysisInput = multiLanguageInput,
+                    Parameters = new DynamicClassificationTaskParameters(
+                        options.DisableServiceLogs,
+                        options.ModelVersion,
+                        options.ClassificationType,
+                        categories.ToList())
+                };
+
+                Response<AnalyzeTextTaskResult> result = _languageRestClient.Analyze(
+                    input,
+                    options.IncludeStatistics,
+                    cancellationToken: cancellationToken);
+
+                DynamicClassificationTaskResult taskResult = (DynamicClassificationTaskResult)result.Value;
+                Response response = result.GetRawResponse();
+
+                IDictionary<string, int> map = CreateIdToIndexMap(multiLanguageInput.Documents);
+                DynamicClassifyDocumentResultCollection results = Transforms.ConvertToDynamicClassifyDocumentResultCollection(taskResult.Results, map);
+                return Response.FromValue(results, response);
+            }
+            catch (Exception e)
+            {
+                scope.Failed(e);
+                throw;
+            }
+        }
+
+        private async Task<Response<DynamicClassifyDocumentResultCollection>> DynamicClassifyBatchAsync(
+            MultiLanguageAnalysisInput multiLanguageInput,
+            IEnumerable<string> categories,
+            DynamicClassifyOptions options,
+            CancellationToken cancellationToken)
+        {
+            options ??= new DynamicClassifyOptions();
+
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(TextAnalyticsClient)}.{nameof(DynamicClassifyBatch)}");
+            scope.Start();
+
+            try
+            {
+                AnalyzeTextDynamicClassificationInput input = new()
+                {
+                    AnalysisInput = multiLanguageInput,
+                    Parameters = new DynamicClassificationTaskParameters(
+                        options.DisableServiceLogs,
+                        options.ModelVersion,
+                        options.ClassificationType,
+                        categories.ToList())
+                };
+
+                Response<AnalyzeTextTaskResult> result = await _languageRestClient.AnalyzeAsync(
+                    input,
+                    options.IncludeStatistics,
+                    cancellationToken: cancellationToken).ConfigureAwait(false);
+
+                DynamicClassificationTaskResult taskResult = (DynamicClassificationTaskResult)result.Value;
+                Response response = result.GetRawResponse();
+
+                IDictionary<string, int> map = CreateIdToIndexMap(multiLanguageInput.Documents);
+                DynamicClassifyDocumentResultCollection results = Transforms.ConvertToDynamicClassifyDocumentResultCollection(taskResult.Results, map);
+                return Response.FromValue(results, response);
             }
             catch (Exception e)
             {